requirements.txt
PyInputPlus>=0.2.12
requests
Pillow
yfinance
tkintermapview